Rename as_json to to_hash
Turns out to_json calls to_hash if its available
(but only on the top-level obj, not recursively)

Also, turns out there is no way to register classes
with stdlib json in order to not have to pollute
my class.

Also, it puts to_json on Object -.-
And its naive in its implementation.