Switch d3 out for jQuery
It wasn't doing anything for me beyond adding the star,
but I can just copy/paste the svg node for the star.

jQuery does everthing I was using it for (I can't tell yet
whether it does a better or worse job of it), and looks like
it can support dom animations, which I couldn't get D3 to do.